pyproject.toml
[tool.poetry]
name = "mlops-fraud-detection"
version = "0.1.0"
description = "Production-grade MLOps pipeline for credit card fraud detection"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# ML & Data
scikit-learn = "^1.4"
xgboost = "^2.0"
lightgbm = "^4.3"
imbalanced-learn = "^0.12"
pandas = "^2.1"
numpy = "^1.26"
optuna = "^3.5"
# MLOps Platform
mlflow = "^2.10"
evidently = "^0.4"
apache-airflow = "^2.8"
# API & Serving
fastapi = "^0.109"
uvicorn = {extras = ["standard"], version = "^0.27"}
pydantic = "^2.5"
pydantic-settings = "^2.1"
# Data Validation
pandera = "^0.18"
# Monitoring
prometheus-client = "^0.19"
# Storage
boto3 = "^1.34"
minio = "^7.2"
# Database
psycopg2-binary = "^2.9"
sqlalchemy = "^2.0"
# Utilities
python-dotenv = "^1.0"
pyyaml = "^6.0"
faker = "^22.0"
# Visualization
matplotlib = "^3.8"
seaborn = "^0.13"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
locust = "^2.20"
# Code Quality
black = "^24.1"
flake8 = "^7.0"
mypy = "^1.8"
isort = "^5.13"
bandit = "^1.7"
pre-commit = "^3.6"
# Jupyter
jupyter = "^1.0"
ipykernel = "^6.29"

[tool.black]
line-length = 100
target-version = ['py310']
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
